02_policy_chatbot/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
streamlit==1.25.0
gradio==3.35.2
openai==0.28.0
langchain==0.0.267
faiss-cpu==1.7.4
jieba==0.42.1
konlpy==0.6.0 